✏️ ♻️ Adding timestamps +rename files in job oppportunity (#49)
* ✏️  Adding timestamp in job opportunity

* ♻️ refactor: Correct file and folder names
