+++ src/application/dtos/jobOpportunity.dto.ts
@@ -8,17 +8,24 @@ export const JobOpportunitySchema = Type.Object({
   salary: Type.Number(),
 });
 
-// Tipos en TypeScript
 export type JobOpportunityDto = Static<typeof JobOpportunitySchema>;
 
-// DTO para creación (sin `id`, ya que lo genera MongoDB)
 export const CreateJobOpportunitySchema = Type.Omit(JobOpportunitySchema, [
   'id',
 ]);
 export type CreateJobOpportunityDto = Static<typeof CreateJobOpportunitySchema>;
 
-// DTO para actualización (todos los campos opcionales)
 export const UpdateJobOpportunitySchema = Type.Partial(
   CreateJobOpportunitySchema,
 );
 export type UpdateJobOpportunityDto = Static<typeof UpdateJobOpportunitySchema>;
+
+export const JobOpportunityResponseSchema = Type.Intersect([
+  JobOpportunitySchema,
+  Type.Object({
+    id: Type.String(),
+    createdAt: Type.String({ format: 'date-time' }),
+    updatedAt: Type.String({ format: 'date-time' }),
+  }),
+]);
+export type MajorResponseDto = Static<typeof JobOpportunityResponseSchema>;
