--- src/application/use-cases/jobOpportunity/delete-jobOpportunity.use-case.ts
@@ -1,11 +1,0 @@
-import { JobOpportunityRepository } from '../../../infrastructure';
-
-export class DeleteJobOpportunityUseCase {
-  constructor(
-    private readonly jobOpportunityRepository: JobOpportunityRepository,
-  ) {}
-
-  public async execute(id: string): Promise<boolean> {
-    return this.jobOpportunityRepository.delete(id);
-  }
-}
