--- src/application/use-cases/jobOpportunity/get-all-jobOpportunity.use-case.ts
@@ -1,12 +1,0 @@
-import { JobOpportunityRepository } from '../../../infrastructure';
-import { JobOpportunity } from '../../../domain';
-
-export class GetAllJobOpportunityUseCase {
-  constructor(
-    private readonly jobOpportunityRepository: JobOpportunityRepository,
-  ) {}
-
-  public async execute(): Promise<JobOpportunity[]> {
-    return this.jobOpportunityRepository.findAll();
-  }
-}
