+++ src/application/use-cases/jobOpportunity/getAllJobOpportunity.useCase.ts
@@ -1,0 +1,12 @@
+import { JobOpportunityRepository } from '../../../infrastructure';
+import { JobOpportunity } from '../../../domain';
+
+export class GetAllJobOpportunityUseCase {
+  constructor(
+    private readonly jobOpportunityRepository: JobOpportunityRepository,
+  ) {}
+
+  public async execute(): Promise<JobOpportunity[]> {
+    return this.jobOpportunityRepository.findAll();
+  }
+}
