+++ src/application/use-cases/jobOpportunity/getJobOpportunityById.useCase.ts
@@ -1,0 +1,12 @@
+import { JobOpportunityRepository } from '../../../infrastructure';
+import { JobOpportunity } from '../../../domain';
+
+export class GetJobOpportunityByIdUseCase {
+  constructor(
+    private readonly jobOpportunityRepository: JobOpportunityRepository,
+  ) {}
+
+  public async execute(id: string): Promise<JobOpportunity | null> {
+    return this.jobOpportunityRepository.findById(id);
+  }
+}
