+++ src/application/use-cases/jobOpportunity/updateJobOpportunity.useCase.ts
@@ -1,0 +1,15 @@
+import { JobOpportunityRepository } from '../../../infrastructure';
+import { JobOpportunity } from '../../../domain';
+
+export class UpdateJobOpportunityUseCase {
+  constructor(
+    private readonly jobOpportunityRepository: JobOpportunityRepository,
+  ) {}
+
+  public async execute(
+    id: string,
+    data: Partial<Omit<JobOpportunity, 'id'>>,
+  ): Promise<JobOpportunity | null> {
+    return this.jobOpportunityRepository.update(id, data);
+  }
+}
