+++ src/infrastructure/database/models/jobOpportunity.model.ts
@@ -5,16 +5,23 @@ interface JobOpportunityDocument extends Document {
   description: string;
   jobId: Types.ObjectId;
   salary: number;
+  createdAt: Date;
+  updatedAt: Date;
 }
 
-const JobOpportunitySchema = new Schema<JobOpportunityDocument>({
-  name: { type: String, required: true },
-  description: { type: String, required: true },
-  jobId: { type: Schema.Types.ObjectId, required: true },
-  salary: { type: Number, required: true },
-});
+const JobOpportunitySchema = new Schema<JobOpportunityDocument>(
+  {
+    name: { type: String, required: true },
+    description: { type: String, required: true },
+    jobId: { type: Schema.Types.ObjectId, required: true },
+    salary: { type: Number, required: true },
+  },
+  {
+    timestamps: true,
+  },
+);
 
 export const JobOpportunityModel = model<JobOpportunityDocument>(
-  'JobOpportunities',
+  'JobOpportunity',
   JobOpportunitySchema,
 );
